+++ CHANGELOG.md
@@ -5,6 +5,8 @@
   * Support parser combinator contexts
   * Rename `literal/2` to `string/2`
   * Add `line/2` and `byte_offset/2` combinators
+  * The MFA given to `repeat_while/3` now receives the accumulator, context, line and byte_offset values
+  * The MFA given to `quoted_repeat_while/3` now receives the accumulator, context, line and byte_offset values
   * The MFA given to `traverse/3` now receives the accumulator, context, line and byte_offset values
   * The MFA given to `quoted_traverse/3` now receives the ASTs for the accumulator, context, line and byte_offset
   * Instead of `line` as a positive integer, we now track `{line, line_offset}` where `line` is the same as before and `line_offset` is the byte_offset after the new line